+++ module/Formation/src/Formation/Service/Url/UrlServiceFactory.php
@@ -2,15 +2,17 @@
 
 namespace Formation\Service\Url;
 
+use Application\RouteMatch;
 use Fichier\Service\Fichier\FichierStorageService;
-use Laminas\View\Renderer\PhpRenderer;
+use Laminas\Router\RouteStackInterface;
 use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\ContainerInterface;
 use Psr\Container\NotFoundExceptionInterface;
 use Structure\Service\StructureDocument\StructureDocumentService;
+use Unicaen\Console\Console;
 
-class UrlServiceFactory {
-
+class UrlServiceFactory
+{
     /**
      * @throws ContainerExceptionInterface
      * @throws NotFoundExceptionInterface
@@ -19,17 +21,23 @@ class UrlServiceFactory {
     {
         /**
          * @var FichierStorageService $fichierStorageService
-         * @var PhpRenderer $renderer
          * @var StructureDocumentService $structureDocumentService
          */
         $fichierStorageService = $container->get(FichierStorageService::class);
-        $renderer = $container->get('ViewRenderer');
         $structureDocumentService = $container->get(StructureDocumentService::class);
 
         $service = new UrlService();
         $service->setFichierStorageService($fichierStorageService);
-        $service->setRenderer($renderer);
         $service->setStructureDocumentService($structureDocumentService);
+
+        /** @var RouteStackInterface $router */
+        $router = $container->get(Console::isConsole() ? 'HttpRouter' : 'Router');
+        $match = $container->get('application')->getMvcEvent()->getRouteMatch();
+        if ($match instanceof RouteMatch) {
+            $service->setRouteMatch($match);
+        }
+        $service->setRouter($router);
+
         return $service;
     }
 }